test: add tests for psycopg3 (#632)
* test: add tests for psycopg3

* test: add more tests

* test: install psycopg3 requirements + cleanup

* test: use format to output byte string

* test: add transaction tests

* test: convert memoryview to bytes

* test: add tests for binary and text formats

* docs: add documentation for psycopg3

* test: add test and documentation for nested transaction

+++ src/main/java/com/google/cloud/spanner/pgadapter/statements/CopyStatement.java
@@ -241,8 +241,8 @@ public class CopyStatement extends IntermediatePortalStatement {
   }
 
   /** @return 0 for text/csv formatting and 1 for binary */
-  public int getFormatCode() {
-    return (parsedCopyStatement.format == Format.BINARY) ? 1 : 0;
+  public byte getFormatCode() {
+    return (parsedCopyStatement.format == Format.BINARY) ? (byte) 1 : (byte) 0;
   }
 
   private void verifyCopyColumns() {
